Bug 1302573: [MSE] P4. Be consistent in datatype being used. r=gerald
When using gcc on linux, about:media shows incorrect content due to incorrect printf formatter.

MozReview-Commit-ID: IWtl6cX1OFA

UltraBlame original commit: 426dfc90db8ad193910e7759f9e98d3b4db4af11

+++ dom/media/mediasource/TrackBuffersManager.h
@@ -308,7 +308,7 @@ private:
     
     
     
-    Maybe<size_t> mNextInsertionIndex;
+    Maybe<uint32_t> mNextInsertionIndex;
     
     TrackBuffer mQueuedSamples;
     
@@ -381,9 +381,9 @@ private:
   
   
   
-  size_t RemoveFrames(const media::TimeIntervals& aIntervals,
-                      TrackData& aTrackData,
-                      uint32_t aStartIndex);
+  uint32_t RemoveFrames(const media::TimeIntervals& aIntervals,
+                        TrackData& aTrackData,
+                        uint32_t aStartIndex);
   
   void ResetEvictionIndex(TrackData& aTrackData);
   void UpdateEvictionIndex(TrackData& aTrackData, uint32_t aCurrentIndex);
@@ -391,7 +391,7 @@ private:
   uint32_t FindSampleIndex(const TrackBuffer& aTrackBuffer,
                            const media::TimeInterval& aInterval);
   const MediaRawData* GetSample(TrackInfo::TrackType aTrack,
-                                size_t aIndex,
+                                uint32_t aIndex,
                                 const media::TimeUnit& aExpectedDts,
                                 const media::TimeUnit& aExpectedPts,
                                 const media::TimeUnit& aFuzz);
